requirements.txt
Cerberus==0.9.2
Eve==0.7.2
Events==0.2.2
Flask==0.12
Flask-Mail==0.9.1
Flask-PyMongo==0.4.1
Jinja2==2.9.6
MarkupSafe==0.23
PyJWT==1.4.2
Werkzeug==0.11.15
argparse==1.2.1
blinker==1.4
click==6.7
itsdangerous==0.24
pycrypto==2.6.1
pymongo==3.4.0
simple-crypt==4.1.7
simplejson==3.10.0
wsgiref==0.1.2
